Improve mouse animation by incorporating velocity (#22)
To improve the smoothness of mouse animation, we want to allow
animations to complete of their own accord. So the question is, how do
we know when the animation is done? The original implementation was, if
you're close to the destination, you're done. This is problematic
because sometimes an animation overshoots the destination, and so the
animation would terminate prematurely. The solution is to take the
derivative of the interpolation which gives the velocity of the
animation. If the position interpolation is near 1 and the velocity
interpolation is near 0, then we infer that the animation is done.